feat: アニメーション管理のロジックを改善し、不要な参照を削除
feat: ビュー切り替え機能を追加し、ページモックアップを実装
