+++ apps/app-nextjs/src/app/(examples)/examples/transitions/_views/view-1.tsx
