+++ apps/app-nextjs/src/app/(examples)/examples/transitions/_views/view-2.tsx
