+++ apps/app-nextjs/src/app/(examples)/examples/transitions/_views/view-3.tsx
