+++ apps/app-nextjs/src/app/(examples)/examples/transitions/page.tsx
@@ -1,8 +1,12 @@
 'use client'
 
 import { TransitionLink } from '@constraints/transitions/components/transition-link'
+import { createFadeStrategy, View, ViewTransitionProvider } from '@/app/_constraints/transitions'
 import { Button } from '@/components/ui/button'
 import { cn } from '@/lib/utils'
+import { View1 } from './_views/view-1'
+import { View2 } from './_views/view-2'
+import { View3 } from './_views/view-3'
 
 interface Props {
   className?: string
@@ -17,6 +21,22 @@ function ViewTransitionsIndex({ className }: Props) {
         </p>
       </div>
       <h1 className="text-3xl font-semibold">Page List</h1>
+
+      <ViewTransitionProvider
+        strategy={createFadeStrategy()}
+        initialViewId="view-1"
+      >
+        <View id="view-1">
+          <View1 />
+        </View>
+        <View id="view-2">
+          <View2 />
+        </View>
+        <View id="view-3">
+          <View3 />
+        </View>
+      </ViewTransitionProvider>
+
       <ul className="flex gap-2">
         <li>
           <Button asChild>
